requirements.txt
click==8.1.7
protobuf==5.26.1
base58==2.1.1
grpcio==1.63.0
grpcio-tools==1.63.0